Separate slow training tests from fast CLI tests
Summary:
Move comprehensive training tests into a dedicated `cli_train_tests.py` file,
keeping `cli_integration_tests.py` fast (one train sanity check + compress/
decompress/flag tests).

- `make test` (via `test-cli`) runs only the fast integration tests.
- `make test-train` runs the full training test suite separately.

Test Plan:
```
make test-cli
```
Runs fast (~5s). All pass.
```
make test-train
```
Runs the comprehensive training tests separately.
